+++ src/common/cpp/common_cpp.cpp
@@ -42,10 +42,10 @@ void clean_up(std::vector<std::string> dirs)
 
 void c_truncate(
     const char* file_path,
-    const int typesz,
-    const int offset,
-    const int movesz,
-    const int buffsz
+    const size_t typesz,
+    const size_t offset,
+    const size_t movesz,
+    const size_t buffsz
 ) {
     if (buffsz <= 0 || typesz <= 0)
     {
